tests init changed
